Add initial version of parser
I've taken the approach that rust-analyzer uses for parsing, so its a
bit of an adventure! So far it's turning out quite nicely, and writing
the parser itself it not too bad. There is, however, a lot of generated
code, and a lot of code to generate that code! The upside to this, is
that I do not think that part will grow that much more then now, and I
think it's a worthy investment!