Rework type system to be Hindley–Milner style
This has quite a large impact on all types in the system, as now one has
to look up the resolved type for each "type id".

The algorithm is a single pass over the AST during lowering from CST to
HIR. If the final type of a type variable is know, that it is set to
that type directly, and if it cannot be determined (like for array
literals []) then the type is set to a new free type.
Later if the type is used in a context where it must satisfy some
constraint, we union the free
type with the expected.