Add monotone analysis, isorecursive pass, and make DB Send+Sync
This is a huge commit with both features and refactoring.

The highlight is the isorecursive pass, which annotates the CFG with the
nessesary folds and unfolds for all field access to be valid.
During this pass, we need information about which slots are live to
maintain access in loop invariants. For this a framework for performing
monotone analysis has been added, and an accompanying liveness analysis.

The liveness analysis fits within the bit-vector/kill-gen framework,
which would be a better fit, but it might be possible to encode the
isorecursive analysis into the monotone framework, so for now we use
monotone as it can serve both purposes.

The isorecursive pipeline itself is not complete: It seems to compute
the correct foldings, but in case of multiple going/outgoing foldings,
merging them is not quite complete. Also the Viper encoding needs some
work.

An unexpected problem is loop conditions. Prior to folding, we could
inline terms in Viper quite liberally, as we did not need to maintain
foldings, and many statements therefore commuted. To play nicely with
folding, most of these inlinings were disabled, which then prevents
loop conditions from being inlined. To solve this, I tried computing the
loop condition immediately before to the loop, and at the very end of
loop iteration. But Viper does not know that the slot placed in the loop
condition is recomputed correctly at each step, so it loops on the
initial value of the conditions. To fix this, an invariant can be added
to the loop,  which establishes the relation between the condition
computation and cond slot. But it is not trivial to find this
computation in the CFG, as excatly what need to be computed is
potentially mangled with other instructions, and it might contain
mutation, which cannot be encoded into an expression required for the
invariant. This problem is still left to be solved.

What has been implemented, is converting all references to the AST in
the HIR to AstPtr, which are Send + Sync, enabling the entire DB to be
so as well! This finally allows us to maintain a single DB instance in
the LSP Backend, and reuse computations, which is (almost) the entrire
purpose of Salsa, so this is nice! One slight hiccup, is that the new
AstPtr's still require a reference the the root of the parsed
source file. We cannot store that, since it will be Send + Sync, but
the GreenNodes would be. The downside of storing these are that they are
quite large in memory, so we need to be careful about how many we store.
For now, we reparse the entire source file everytime we need access to
the root, and luckly the performance does not seem to be too bad.