Change the type of return expressions to be free instead of error
Semantically these are kind of the similar: both get unified with what
ever they meet, but error propagates the error (mostly) type, while free
takes the appropriate type.

Thinking about it, perhaps we should rename the free type data variant
to never, being the bottom type, and let error be the top type. We would
still have free in terms of type ids, but those all resolve to the never
type which is unifiable with what ever it needs to be.