Refactor names in HIR
Instead of using `Ident`, which also carried a span along, we not use a
`Name` which simply wraps `SmolStr`. To get spans the, one need to find
an `AstPtr<..>` to what ever the wrapping type is, and navigate to the
desired string.

This is a step in the direction of not having spans in the HIR and lower
levels of the compiler, and instead relying on `AstPtr` and source maps.