Gen while decreases, mangle names, refactor folding
This commit started out by implementing lowering of decreases in while
loops, but it turned out quite a few bugs had snuck in during the last
refactoring of the folding tree.

The biggest change made, is for monotone framework to compute facts for
every body location, instead of only at the entry/exit of blocks.
This aligns it better with the theory, so that is a nother plus.
But the transition to this new structure was not super trivial.

In the end, I hope the code has gotten easier to debug, and hopefully
more correct.

In addition to this, a trait for a MIR pass was added, where we now have
two passes:

- The mention pass, which makes sures that roots of structs are
  mentioned in loop heads, so that their access is preserved.
- The isorecursive pass, which computes and inserts foldings in the MIR.