Restructure typing layout in crate
This introduces a new top module in mist_core, `types`, which contains
almost everything related to types and typing. Notably it does not
contain code for actual type checking, but rather the low-level
primitives for doing type unification and looking up the data of a
type id.

Another big change, is that `TypeData` has been split into two types:
- The old `TypeData` enum is now `TypeDataKind` aka `TDK`.
- `Ghost` has been removed from `TDK`, and `TypeData` is now a struct
  with a kind, and an `is_ghost` flag.

This simplifies a lot of matching, where one previously had to do
`.strip_ghost()` before looking at the variant. The downside is that one
must be careful and remember to propagate the ghost flag when
transforming types.

Further, a file context has been added, which can compute the types and
functions once per file, instead of once per definition as done before.
This is one step in the direction of simplifying initialization of the
typechecker, and in general the type checking.