Track NameOrNameRef in sourcemap and remove VariableId and VariableRef
This simplifies a lot of code regarding variables. Instead of tracking
spans in every ref, we now lazily compute them from the AST.

The visitor also uses the rowan preorder now to find all names, and then
visiting them using the source map. In the future, more visiting should
be done like this.