Implement struct initialization
We had to move initialization from MExpr to a separate instruction,
since constructing access is not allowed in pure contexts in Viper.

This also means that trying to lower a struct initialization in a pure
context results in an error.

Potential solutions to lowering in pure contexts are doing special
"pure domain" versions of structs, but that is a whole langauge feature
I think. TBD.