Improve folding analysis of field assignment
Previously we droped everything being assigned to, which resulted in
some things unnecessarily loosing access to fields. Instead we now still
perform the drop, but then do a soft-require on the field.

Soft-require is a new term, which does the same as require, but does
not force the soft-required field to be necessarily be folded, rather
it should _just_ be accessible. This is exactly the requirement for
performing assignments, where the concrete foldings of the field and its
children are not important.