Add failing list of structs example
When trying to access a field of a struct inside a list, the current
backwards analysis for foldings propagates the unfolding to a point
before the index has been computed.

This results in the unfolding being performed while the index is
uninitialized.

I don't know if the solution to this yet. Potentially we could delay
unfoldings for as long as possible, but that does not immediately fit
with the current backwards constraint solving approach.