Refactor Structs to be ADT and start grouping on items instead of defs
This is a very inprogress commit.

Currently invariants do not get lowered, and mir lowering is in an
inbetween state.

The next big thing, is introducing a new stage: Mono.

The mono stage accumulates the definitions from HIR, and creates a
consistent view over the HIR. This, most importantly, includes interning
the types, such that they are the same if they contain the same
TypeData. This turns out to be important when merging multiple HIR
definitions; something that becomes apparent with generic ADT
instantiation.