Split signature related stuff out of Body into ItemBody
This commit also contains a huge refactor of the way this in MIR is
accessed. For example, place now holds a type, so you won't have to look
it up in the body. The same thing would be nice for slot, but I've not
figured out how to do it just yet.