Add pure structs
Pure structs are structs annotated with pure. Their main difference
is the way they are lowered, and their semantics.

Comparing pure structs to Rust concepts, they are structs with
equality and copy. This inherintly means that they cannot be
recursive, and must be finite in size.

In Viper, pure structs are lowered as domains, rather than
predicates. This also means that they are not subject to folding
and unfolding, and their invariants are described using domain
axioms, and field access is through domain functions. To make sure
that constructed pure structs satisfy their invariants, creating one
must go through an abstract function, which checks the invariant as a
precondition. This final part is still not quite perfectly
implemented.

Additionally, this commit fixes some folding-forrest tests, and
begins work for testing with fixtures.