Attemp to cancel verification if invalidated
It is currently not quite documented in salsa how you are supposed to
correctly handle cancelation, so we try to do the best we can.

By inserting probes to a dummy query, we give the salsa runtime a chance
to interupt the running query if invalidated. A comment has been added
that this should be redone when it is more clear how this is best done
in salsa.