Introduce `let mut` which will be ghost in pure
The checking is not complete in all cases, but it does allow usage of
`mut` in pure funcitons by stripping them out in the final function.

Impure code in pure functions is generally useful for using lemmas and/
or asserting properties.

Managing what code gets emitted were is crucial for this to work, and
statically checking that these constraints hold is also important!

More work left to be done!